[STRUCTURAL] Extract Class: skillPromptContributor from prompt build
buildSystemPromptWithFragments carried ~130 lines of skill-runtime
integration inline — four local types, two closures, fragment selection,
and the before-prompt-build hook mutation protocol — coupling the agent's
central prompt-build method to the skill hook wire format in fine detail.

That logic moves into skillPromptContributor (internal/agent/skill_prompt.go),
which selects budgeted fragments, dispatches HookOnBeforePromptBuild, and
applies the requested mutations, returning the resolved base prompt and
fragments. buildSystemPromptWithFragments drops from ~180 lines to ~39: a
thin orchestrator that lets the contributor speak, assembles static +
strategy sections, then renders fragments.

The contributor is deliberately NOT an agentsdk.ContextStrategy: a hook can
replace the base system prompt wholesale, a whole-prompt transform the
section-contribution seam (append-only) cannot express — documented on the
type. Behavior preserved; the existing skill/hook suites plus the
replace_base_system_prompt pin stay green.
